requirements.txt
coverage==4.5.1
flake8==3.5.0
GitPython==2.1.11
green==2.12.1
mock==2.0.0
pydocstyle==2.1.1
pylint==2.1.1
sphinx-argparse==0.2.2
sphinx-rtd-theme==0.4.1
Sphinx==1.7.6
pyyaml==3.13
